Widen price fields from int to long to eliminate overflow at ~2.1B
currentPrice/basePrice/listingPrice in Stock entity were int, causing
silent negative overflow when price exceeded Integer.MAX_VALUE (~2.1B).
Changed to long (max ~9.2×10^18) and added V51 DB migration.

+++ backend/src/main/java/com/spotchzxk/repository/StockRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface StockRepository extends JpaRepository<Stock, String> {
     List<Stock> findByIsLiveTrue();
-    List<Stock> findByCurrentPriceGreaterThan(int currentPrice);
+    List<Stock> findByCurrentPriceGreaterThan(long currentPrice);
 }
